requirements.txt
streamlit==1.47.0
opencv-python-headless==4.7.0.72
numpy==1.23.5
tensorflow==2.12.0
Pillow==11.3.0
pandas==2.3.1
matplotlib==3.10.3
streamlit-webrtc